requirements.txt
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
<<<<<<< HEAD
Django==5.1.6
django-debug-toolbar==5.0.1
django-extensions==3.2.3
h11==0.14.0
=======
Django==5.1.7
django-debug-toolbar==5.0.1
django-extensions==3.2.3
h11==0.14.0
html5lib==1.1
>>>>>>> recuperacion
idna==3.10
lxml==5.3.1
mysqlclient==2.2.7
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
<<<<<<< HEAD
scripts==3.0
=======
>>>>>>> recuperacion
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
<<<<<<< HEAD
stua==0.3
=======
>>>>>>> recuperacion
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
<<<<<<< HEAD
urllib3==2.3.0
webdriver-manager==4.0.2
=======
Unidecode==1.3.8
urllib3==2.3.0
webdriver-manager==4.0.2
webencodings==0.5.1
>>>>>>> recuperacion
websocket-client==1.8.0
wsproto==1.2.0
